+++ SourceCode/App/Validators/ModuleValidator.cs
@@ -17,10 +17,18 @@ namespace ModulesRegistry.Validators
             RuleFor(m => m.ScaleId)
                 .MustBeSelected(localizer)
                 .WithName(n => localizer[nameof(n.Scale)]);
-            RuleFor(m => m.FREMONumber)
-                .InclusiveBetween(1,9999)
-                .When(m => m.FREMONumber.HasValue)
-                .WithName(n => localizer[nameof(n.FREMONumber)]);
+            RuleFor(m => m.ConfigurationLabel)
+                .MaximumLength(10)
+                .When(m => m.ConfigurationLabel is not null)
+                .WithName(n => localizer[nameof(n.ConfigurationLabel)]);
+            RuleFor(m => m.PackageLabel)
+                .MaximumLength(10)
+                .When(m => m.PackageLabel is not null)
+                .WithName(n => localizer[nameof(n.PackageLabel)]);
+            RuleFor(m => m.FremoNumber)
+                .InclusiveBetween(1, 9999)
+                .When(m => m.FremoNumber.HasValue)
+                .WithName(n => localizer[nameof(n.FremoNumber)]);
             RuleFor(m => m.StandardId)
                 .MustBeSelected(localizer)
                 .WithName(n => localizer[nameof(n.Standard)]);
@@ -31,7 +39,7 @@ namespace ModulesRegistry.Validators
                .MustBeValidYear(localizer)
                .WithName(n => localizer["UptoYear"]);
             RuleFor(m => m.Radius)
-                .InclusiveBetween(0.0, 10.0)
+                .InclusiveBetween(0.0, 10000.0)
                 .When(m => m.Radius is not null)
                 .WithName(n => localizer[nameof(n.Radius)]);
             RuleFor(m => m.Angle)
@@ -39,16 +47,16 @@ namespace ModulesRegistry.Validators
                 .When(m => m.Angle is not null)
                 .WithName(n => localizer[nameof(n.Angle)]);
             RuleFor(m => m.Length)
-                .InclusiveBetween(0, 50)
+                .InclusiveBetween(5, 50000)
                 .WithName(n => localizer[nameof(n.Length)]);
             RuleFor(m => (int)m.NumberOfThroughTracks)
                 .InclusiveBetween(1, 4)
                 .WithName(n => localizer[nameof(n.NumberOfThroughTracks)]);
             RuleFor(m => m.FunctionalState)
-                .IsInEnum()
+                .InclusiveBetween((int)ModuleFunctionalState.Unknown, (int)ModuleFunctionalState.Approved)
                 .WithName(n => localizer[nameof(n.FunctionalState)]);
             RuleFor(m => m.LandscapeState)
-                .IsInEnum()
+                .InclusiveBetween((int)ModuleLandscapeState.Unknown, (int)ModuleLandscapeState.FullyAppliedDetailed)
                 .WithName(n => localizer[nameof(n.LandscapeState)]);
         }
     }