+++ SourceCode/Data/User.cs
@@ -21,5 +21,6 @@ namespace ModulesRegistry.Data
         public bool IsGlobalAdministrator { get; set; }
         public bool IsCountryAdministrator { get; set; }
         public string HashedPassword { get; set; }
+
     }
 }